+++ docs/asciidoc/modules/avaje-inject.adoc
@@ -21,7 +21,7 @@
           <path>
             <groupId>io.avaje</groupId>
             <artifactId>avaje-inject-generator</artifactId>
-            <version>10.0</version>
+            <version>10.3</version>
           </path>
         </annotationProcessorPaths>
       </configuration>
