+++ modules/jooby-avaje-validator/pom.xml
@@ -11,7 +11,11 @@
 
   <modelVersion>4.0.0</modelVersion>
   <artifactId>jooby-avaje-validator</artifactId>
-
+  
+  <properties>
+    <maven.compiler.proc>full</maven.compiler.proc>
+  </properties>
+  
   <dependencies>
     <dependency>
       <groupId>io.jooby</groupId>
@@ -41,7 +45,18 @@
       <artifactId>jooby-netty</artifactId>
       <scope>test</scope>
     </dependency>
-    
+
+    <dependency>
+      <groupId>io.jooby</groupId>
+      <artifactId>jooby-apt</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.avaje</groupId>
+      <artifactId>avaje-validator-generator</artifactId>
+      <scope>test</scope>
+    </dependency>
+
     <dependency>
       <groupId>io.jooby</groupId>
       <artifactId>jooby-jackson</artifactId>
@@ -95,16 +110,6 @@
           <compilerArgs>
             <arg>-parameters</arg>
           </compilerArgs>
-          <annotationProcessorPaths>
-            <path>
-              <groupId>io.jooby</groupId>
-              <artifactId>jooby-apt</artifactId>
-            </path>   
-            <path>
-              <groupId>io.avaje</groupId>
-              <artifactId>avaje-validator-generator</artifactId>
-             </path>
-          </annotationProcessorPaths>
         </configuration>
       </plugin>
     </plugins>
