+++ modules/jooby-avaje-validator/src/test/java/io/jooby/avaje/validator/app/PasswordsShouldMatch.java
@@ -1,4 +1,4 @@
-package io.jooby.hibernate.validator.app;
+package io.jooby.avaje.validator.app;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -10,13 +10,11 @@ import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-@Constraint(validatedBy = PasswordsShouldMatchValidator.class)
+@Constraint(validatedBy = {})
 @Target({TYPE, ANNOTATION_TYPE})
 @Retention(RUNTIME)
 public @interface PasswordsShouldMatch {
-    String message() default "Passwords should match";
+  String message() default "Passwords should match";
 
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
+  Class<?>[] groups() default {};
 }
