+++ modules/jooby-avaje-validator/src/test/java/io/jooby/avaje/validator/app/PasswordsShouldMatchValidator.java
@@ -1,0 +1,21 @@
+package io.jooby.avaje.validator.app;
+
+import io.avaje.validation.adapter.AbstractConstraintAdapter;
+import io.avaje.validation.adapter.ConstraintAdapter;
+import io.avaje.validation.adapter.ValidationContext.AdapterCreateRequest;
+
+@ConstraintAdapter(PasswordsShouldMatch.class)
+public class PasswordsShouldMatchValidator extends AbstractConstraintAdapter<NewAccountRequest> {
+
+  public PasswordsShouldMatchValidator(AdapterCreateRequest request) {
+    super(request);
+  }
+
+  @Override
+  public boolean isValid(NewAccountRequest request) {
+    if (request.getPassword() == null || request.getConfirmPassword() == null) {
+      return false;
+    }
+    return request.getPassword().equals(request.getConfirmPassword());
+  }
+}
