--- modules/jooby-avaje-validator/src/test/java/io/jooby/hibernate/validator/app/NewAccountRequest.java
@@ -1,59 +1,0 @@
-package io.jooby.hibernate.validator.app;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-
-@PasswordsShouldMatch
-public class NewAccountRequest {
-    @NotNull
-    @NotEmpty
-    @Size(min = 3, max = 16)
-    private String login;
-
-    @NotNull
-    @NotEmpty
-    @Size(min = 8, max = 24)
-    private String password;
-
-    @NotNull
-    @NotEmpty
-    @Size(min = 8, max = 24)
-    private String confirmPassword;
-
-    @Valid
-    private Person person;
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getConfirmPassword() {
-        return confirmPassword;
-    }
-
-    public void setConfirmPassword(String confirmPassword) {
-        this.confirmPassword = confirmPassword;
-    }
-
-    public Person getPerson() {
-        return person;
-    }
-
-    public void setPerson(Person person) {
-        this.person = person;
-    }
-}
