--- modules/jooby-avaje-validator/src/test/java/io/jooby/hibernate/validator/app/PasswordsShouldMatchValidator.java
@@ -1,15 +1,0 @@
-package io.jooby.hibernate.validator.app;
-
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-public class PasswordsShouldMatchValidator implements ConstraintValidator<PasswordsShouldMatch, NewAccountRequest> {
-
-    @Override
-    public boolean isValid(NewAccountRequest request, ConstraintValidatorContext constraintContext) {
-        if (request.getPassword() == null || request.getConfirmPassword() == null) {
-            return false;
-        }
-        return request.getPassword().equals(request.getConfirmPassword());
-    }
-}
