--- modules/jooby-avaje-validator/src/test/java/io/jooby/hibernate/validator/app/Person.java
@@ -1,35 +1,0 @@
-package io.jooby.hibernate.validator.app;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-
-@Valid
-public class Person {
-
-    @NotEmpty
-    @NotNull
-    private String firstName;
-    private String lastName;
-
-    public Person(String firstName, String lastName) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-}
