+++ modules/jooby-validation/src/main/java/io/jooby/validation/MvcValidator.java
@@ -1,5 +1,7 @@
 package io.jooby.validation;
 
+import io.jooby.Context;
+
 /**
  * This interface should be implemented by modules that provide bean validation functionality.
  * An instance of this interface must be registered in the Jooby service registry.
@@ -11,7 +13,8 @@ public interface MvcValidator {
     /**
      * Method should validate the bean and throw an exception if any constraint violations are detected
      * @param bean bean to be validated
+     * @param ctx request context
      * @throws RuntimeException an exception with violations to be thrown (e.g. ConstraintViolationException)
      */
-    void validate(Object bean) throws RuntimeException;
+    void validate(Context ctx, Object bean) throws RuntimeException;
 }
