+++ modules/pom.xml
@@ -60,6 +60,7 @@
 
     <!-- validation -->
     <module>jooby-validation</module>
+    <module>jooby-avaje-validator</module>
     <module>jooby-hibernate-validator</module>
 
     <module>jooby-pac4j</module>
